Revert "Update Better Game Menu Integration"
This reverts commit f1482a96fefda6282d100ee4e2423d8a42de9989.
