+++ SocialPageOrderMenu/ModEntry.cs
@@ -19,7 +19,7 @@ using SocialPageOrderRedux.UI;
 
 namespace SocialPageOrderRedux
 {
-    public partial class ModEntry : Mod
+    public class ModEntry : Mod
     {
         public static ModConfig Config;
         public static IMonitor SMonitor;
@@ -133,11 +133,6 @@ namespace SocialPageOrderRedux
             helper.Events.Display.MenuChanged += Display_MenuChanged;
             helper.Events.Content.LocaleChanged += Content_LocaleChanged;
 
-            if(SHelper.ModRegistry.IsLoaded(IDs.BetterGameMenu))
-            {
-                return;
-            }
-
             var harmony = new Harmony(ModManifest.UniqueID);
 
             harmony.Patch(AccessTools.Constructor(typeof(SocialPage), new Type[] { typeof(int), typeof(int), typeof(int), typeof(int) }),
@@ -202,7 +197,16 @@ namespace SocialPageOrderRedux
 
             if(BetterGameMenuAPI is not null)
             {
-                Setup(BetterGameMenuAPI);
+                BetterGameMenuAPI.OnTabChanged(BetterGameMenuChangeTab);
+                BetterGameMenuAPI.OnMenuCreated(BetterGameMenuMenuCreated);
+
+                // Add the receiveLeftClick patches to BetterGameMenu. These just set a flag when the method is entered and left that makes it so the filter being selected does not make readyToClose on the SocialPage be false.
+                // This is necessary because BetterGameMenu (and the base game) checks readyToClose before changing tabs, so without this it would not be possible to click on another tab when the filter is selected.
+                var harmony = new Harmony(ModManifest.UniqueID);
+                harmony.Patch(AccessTools.Method(BetterGameMenuAPI.GetMenuType(), nameof(IClickableMenu.receiveLeftClick)), // The Better Game Menu left click has necessarily the same name as IClickableMenu.receiveLeftClick
+                    prefix: new HarmonyMethod(typeof(ModEntry), nameof(GameMenu_receiveLeftClick_Prefix)),
+                    postfix: new HarmonyMethod(typeof(ModEntry), nameof(GameMenu_receiveLeftClick_Postfix))
+                );
             }
 
             // get Generic Mod Config Menu's API (if it's installed)
@@ -432,6 +436,21 @@ namespace SocialPageOrderRedux
 
         #region SOCIAL_PAGE_SETUP_PATCHES
 
+        /// <summary>
+        /// Prevents a null error due to the order that BetterGameMenu events fire relative to our patches.
+        /// </summary>
+        public static void BetterGameMenuMenuCreated(IClickableMenu menu)
+        {
+            try
+            {
+                InitElements();
+            }
+            catch (Exception ex)
+            {
+                SMonitor.Log($"Failed in {nameof(BetterGameMenuMenuCreated)}: {ex}", LogLevel.Error);
+            }
+        }
+
         public static void SocialPage_Constructor_Postfix(SocialPage __instance)
         {
             if (!Config.EnableMod)
@@ -568,7 +587,7 @@ namespace SocialPageOrderRedux
         public static void IClickableMenu_readyToClose_Postfix(IClickableMenu __instance, ref bool __result)
         {
             // isInGameMenuLeftClick is necessary so that the player doesn't need to deselect the field before moving to another tab
-            if (!Config.EnableMod || __instance is not SocialPage || filterField.Value is null || isInGameMenuLeftClick.Value || BetterGameMenuAPI is not null)
+            if (!Config.EnableMod || __instance is not SocialPage || filterField.Value is null || isInGameMenuLeftClick.Value)
                 return;
 
             // If the filter is selected, make the result false. This is because some mods that add their own menus will overwrite the current one when their menu's key is pressed,
@@ -715,6 +734,35 @@ namespace SocialPageOrderRedux
                 filterField.Value.Selected = false;
             }
         }
+
+        public static void BetterGameMenuChangeTab(ITabChangedEvent e)
+        {
+            try
+            {
+                if (BetterGameMenuAPI.ActivePage is SocialPage page)
+                {
+                    if (Config.UseFilter && filterField.Value is not null && !Game1.options.gamepadControls)
+                        filterField.Value.Selected = Config.SearchBarAutoFocus;
+
+                    // I don't think this is necessary, but I'm not removing it because it also isn't hurting anything.
+                    ApplyFilter(page);
+                    if (Game1.options.SnappyMenus)
+                    {
+                        page.snapToDefaultClickableComponent();
+                    }
+                }
+                else
+                {
+                    filterField.Value.Selected = false;
+                }
+            }
+            catch (Exception ex)
+            {
+                SMonitor.Log($"Failed in {nameof(BetterGameMenuChangeTab)}: {ex}", LogLevel.Error);
+            }
+
+        }
+
         #endregion
 
         #region SORT_AND_FILTER_METHODS
@@ -1020,7 +1068,7 @@ namespace SocialPageOrderRedux
             return null;
         }
 
-        public static void UpdateElementPositions(SocialPage page)
+        private static void UpdateElementPositions(SocialPage page)
         {
             if(Config.UseButton)
             {
